+++ src/main/java/roomescape/common/exception/ExceptionApiController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import roomescape.reservation.controller.dto.response.PaymentErrorResponse;
+import roomescape.payment.dto.resonse.PaymentErrorResponse;
 
 @RestControllerAdvice(annotations = RestController.class)
 public class ExceptionApiController {
