+++ src/main/java/roomescape/payment/dto/resonse/PaymentErrorResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.controller.dto.response;
+package roomescape.payment.dto.resonse;
 
 public record PaymentErrorResponse(
         String message
