+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -4,6 +4,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.auth.dto.LoginMember;
+import roomescape.payment.dto.request.PaymentConfirmRequest;
+import roomescape.payment.service.PaymentService;
 import roomescape.reservation.controller.dto.request.ReservationSaveRequest;
 import roomescape.reservation.controller.dto.request.ReservationSearchCondRequest;
 import roomescape.reservation.controller.dto.response.MemberReservationResponse;
@@ -11,7 +13,6 @@ import roomescape.reservation.controller.dto.response.ReservationResponse;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.Status;
 import roomescape.reservation.repository.ReservationRepository;
-import roomescape.reservation.service.dto.request.PaymentConfirmRequest;
 import roomescape.reservation.service.dto.request.ReservationPaymentRequest;
 
 @Service
