+++ src/test/java/roomescape/reservation/repository/ReservationRepositoryTest.java
@@ -30,6 +30,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberName;
 import roomescape.member.repository.MemberRepository;
+import roomescape.reservation.controller.dto.request.ReservationSearchCondRequest;
 import roomescape.reservation.domain.Description;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
@@ -37,7 +38,6 @@ import roomescape.reservation.domain.ReservationWithRank;
 import roomescape.reservation.domain.Status;
 import roomescape.reservation.domain.Theme;
 import roomescape.reservation.domain.ThemeName;
-import roomescape.reservation.controller.dto.request.ReservationSearchCondRequest;
 
 @DataJpaTest
 class ReservationRepositoryTest {
