+++ src/test/java/roomescape/reservation/service/ReservationSaveServiceTest.java
@@ -19,6 +19,7 @@ import org.springframework.test.context.ActiveProfiles;
 import roomescape.common.config.DatabaseCleaner;
 import roomescape.member.domain.Member;
 import roomescape.member.repository.MemberRepository;
+import roomescape.payment.service.PaymentService;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.Theme;
 import roomescape.reservation.repository.ReservationTimeRepository;
